# RadiKeep/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RadiKeep | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://127.0.0.1:7262/"
    }
  ],
  "paths": {
...
    "/api/settings/external-import/scan": {
      "post": {
        "tags": [
          "ExternalImportApi"
        ],
        "summary": "録音保存先をスキャンして未登録候補を取得する",
        "operationId": "ApiExternalImportScan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ExternalImportScanRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfListOfExternalImportCandidateEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/external-import/export-csv": {
      "post": {
        "tags": [
          "ExternalImportApi"
        ],
        "summary": "候補一覧をCSVでダウンロードする",
        "operationId": "ApiExternalImportExportCsv",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalImportCandidatesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/settings/external-import/import-csv": {
      "post": {
        "tags": [
          "ExternalImportApi"
        ],
        "summary": "CSVをアップロードして候補一覧を再構築する",
        "operationId": "ApiExternalImportImportCsv",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfListOfExternalImportCandidateEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/external-import/save": {
      "post": {
        "tags": [
          "ExternalImportApi"
        ],
        "summary": "候補を録音済み番組へ保存する",
        "operationId": "ApiExternalImportSave",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalImportCandidatesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfExternalImportSaveResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfExternalImportSaveResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/external-import/maintenance/scan-missing": {
      "post": {
        "tags": [
          "ExternalImportApi"
        ],
        "summary": "録音ファイル欠損レコードをスキャンする",
        "operationId": "ApiExternalImportScanMissing",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfRecordingFileMaintenanceScanResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/external-import/maintenance/relink-missing": {
      "post": {
        "tags": [
          "ExternalImportApi"
        ],
        "summary": "欠損レコードのファイル再紐付けを実行する",
        "operationId": "ApiExternalImportRelinkMissing",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingMaintenanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfRecordingFileMaintenanceActionResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/external-import/maintenance/delete-missing": {
      "post": {
        "tags": [
          "ExternalImportApi"
        ],
        "summary": "欠損レコードをDBから削除する",
        "operationId": "ApiExternalImportDeleteMissing",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingMaintenanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfRecordingFileMaintenanceActionResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseOfEmptyData": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EmptyData"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ApiResponseOfExternalImportSaveResult": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExternalImportSaveResult"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfListOfExternalImportCandidateEntry": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ExternalImportCandidateEntry"
            }
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfRecordingFileMaintenanceActionResult": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RecordingFileMaintenanceActionResult"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ApiResponseOfRecordingFileMaintenanceScanResult": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RecordingFileMaintenanceScanResult"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "EmptyData": {
        "type": "object"
      },
      "ExternalImportCandidateEntry": {
        "type": "object",
        "properties": {
          "isSelected": {
            "type": "boolean"
          },
          "filePath": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "stationName": {
            "type": "string"
          },
          "broadcastAt": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ExternalImportCandidatesRequest": {
        "type": "object",
        "properties": {
          "candidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalImportCandidateEntry"
            }
          },
          "markAsListened": {
            "type": "boolean"
          }
        }
      },
      "ExternalImportSaveResult": {
        "type": "object",
        "properties": {
          "savedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalImportValidationError"
            }
          }
        }
      },
      "ExternalImportScanRequest": {
        "type": "object",
        "properties": {
          "applyDefaultTag": {
            "type": "boolean"
          }
        }
      },
      "ExternalImportValidationError": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "RecordingFileMaintenanceActionDetail": {
        "type": "object",
        "properties": {
          "recordingId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "RecordingFileMaintenanceActionResult": {
        "type": "object",
        "properties": {
          "targetCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "successCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "skipCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "failCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordingFileMaintenanceActionDetail"
            }
          }
        }
      },
      "RecordingFileMaintenanceEntry": {
        "type": "object",
        "properties": {
          "recordingId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "stationName": {
            "type": "string"
          },
          "storedPath": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "candidateCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "candidateRelativePaths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RecordingFileMaintenanceScanResult": {
        "type": "object",
        "properties": {
          "missingCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordingFileMaintenanceEntry"
            }
          }
        }
      },
      "RecordingMaintenanceRequest": {
        "type": "object",
        "properties": {
          "recordingIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}